Change LastPlayed to int type
This value is a unix timestamp so normally it couldbe parsed straight into a DateTime object, but if the hero has never been played by the player then the value is 0 which would break the DateTime parsing. For now the LastPlayed data is not used anywhere so it can be kept as an int to handle any of these cases.